:recycle: [REFACTOR] queryDsl parameter : Long -> Integer

+++ backend/nuguri/src/main/java/com/ssafy/nuguri/dto/hobby/HobbyHistoryResponseDto.java
@@ -26,7 +26,7 @@ public class HobbyHistoryResponseDto {
 
     private Integer maxNum;
 
-    private Long wishlistNum;
+    private Integer wishlistNum;
 
     private Integer chatNum;
